requirements.txt
flask==3.0.0
selenium==4.16.0
beautifulsoup4==4.12.2
tinydb==4.8.0
pandas==2.2.0
plotly==5.18.0
kaleido==0.2.1
scipy==1.12.0
matplotlib==3.8.2
numpy==1.26.3
requests==2.31.0
python-dotenv==1.0.1
pyarrow==15.0.0
